move TestRail API client into separate testrail_api package, and include the software license
